+++ packages/ipc/src/ipc.ts
@@ -39,7 +39,13 @@ function ipc({ socketTimeout = 30000, path = '/tmp/app.socket-' }: IPCConfig = {
 
     function connect(handle: net.Socket): net.Socket {
       const connection: net.Socket = handle || net.connect(`${path}${params.port}`);
-      connection.on('data', req => emitter.emit('rawFrame', req));
+      connection.on('data', req => {
+        const chunks = req.toString().split('\n\n');
+        for (let i = 0; i < chunks.length; i++) {
+          if (chunks[i][0] !== '{' || chunks[i][chunks[i].length - 1] !== '}') continue;
+          emitter.emit('frame', JSON.parse(chunks[i]), req.length);
+        }
+      });
       connection.on('error', err => emitter.emit('error', err));
       connection.on('connect', () => emitter.emit('connect', connection));
       connection.on('close', () => emitter.emit('disconnect'));
@@ -51,8 +57,8 @@ function ipc({ socketTimeout = 30000, path = '/tmp/app.socket-' }: IPCConfig = {
       if (listener) listener.close();
     }
 
-    function send(handle: net.Socket, payload: number[]): void {
-      if (handle) handle.write(Buffer.from(payload));
+    function send(handle: net.Socket, payload: RawFrame): void {
+      if (handle) handle.write(`${JSON.stringify(payload)}\n\n`);
     }
 
     return {