+++ packages/kalm/src/routines/realtime.ts
@@ -1,19 +1,16 @@
 /* Methods -------------------------------------------------------------------*/
 
 export function realtime(): KalmRoutine {
-  return function queue(channel: string, params: object, channelEmitter: NodeJS.EventEmitter, clientEmitter: NodeJS.EventEmitter): Queue {
-    let i: number = 0;
+  return function queue(params: object, routineEmitter: NodeJS.EventEmitter): Queue {
+    let frameId: number = 0;
 
-    function add(packet: Buffer): void {
-      channelEmitter.emit('runQueue', { frameId: i++, channel, packets: [packet] });
-      if (i > 255) i = 0;
-      clientEmitter.emit(`${channel}.queueAdd`, { frameId: i, packet: 0 });
-      clientEmitter.emit(`${channel}.queueRun`, { frameId: i, packets: 1 });
+    function add(): void {
+      routineEmitter.emit('runQueue', { frameId });
+      if (++frameId > 0xffff) frameId = 0;
     }
 
-    function size(): number { return 0; }
     function flush(): void {}
 
-    return { add, size, flush };
+    return { add, flush, emitter: routineEmitter };
   };
 }